- silence gcc liberal enum value handling
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code/trunk/gwyddion@11602 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ modules/file/rawxyz.c
@@ -1024,7 +1024,7 @@ rawxyz_do(RawXYZFile *rfile,
         fill_field_x((const GwyTriangulationPointXYZ*)points->data, dfield);
     else if (rfile->regular == RAW_XYZ_REGULAR_Y)
         fill_field_y((const GwyTriangulationPointXYZ*)points->data, dfield);
-    else if (args->interpolation == GWY_INTERPOLATION_FIELD) {
+    else if ((gint)args->interpolation == GWY_INTERPOLATION_FIELD) {
         extend_borders(rfile, args, EPSREL);
         interpolate_field(points->len,
                           (const GwyTriangulationPointXYZ*)points->data,
@@ -1709,7 +1709,7 @@ static void
 rawxyz_sanitize_args(RawXYZArgs *args)
 {
     if (args->interpolation != GWY_INTERPOLATION_ROUND
-        && args->interpolation != GWY_INTERPOLATION_FIELD)
+        && (gint)args->interpolation != GWY_INTERPOLATION_FIELD)
         args->interpolation = GWY_INTERPOLATION_LINEAR;
     if (args->exterior != GWY_EXTERIOR_MIRROR_EXTEND
         && args->exterior != GWY_EXTERIOR_PERIODIC)
